refactor: Game, User, Random 클래스 추가 Game: 게임 핵심 로직을 담당하는 클래스 User: 유저 입력 담당하는 클래스 Random: 랜덤값을 담당하는 클래스
